move deep merge to dict_utils
