+++ vivarium/actor/process.py
@@ -11,7 +11,7 @@ import matplotlib.pyplot as plt
 
 import vivarium.actor.emitter as emit
 from vivarium.utils.units import units
-from vivarium.utils.dict_utils import merge_dicts
+from vivarium.utils.dict_utils import merge_dicts, deep_merge
 
 
 
@@ -196,19 +196,6 @@ def merge_default_updaters(processes):
         default_updaters = settings['updaters']
         updaters = deep_merge(dict(updaters), default_updaters)
     return updaters
-
-def deep_merge(dct, merge_dct):
-    '''
-    Recursive dict merge
-    This mutates dct - the contents of merge_dct are added to dct (which is also returned).
-    If you want to keep dct you could call it like deep_merge(dict(dct), merge_dct)'''
-    for k, v in merge_dct.items():
-        if (k in dct and isinstance(dct[k], dict)
-                and isinstance(merge_dct[k], collections.Mapping)):
-            deep_merge(dct[k], merge_dct[k])
-        else:
-            dct[k] = merge_dct[k]
-    return dct
 
 def default_divide_condition(compartment):
     return False
@@ -580,7 +567,6 @@ def simulate_process_with_environment(process, settings={}):
     topology = {
         'process': {key: key for key in states},
     }
-
 
     options = {
         'topology': topology,
