+++ vivarium/processes/metabolism.py
@@ -7,10 +7,10 @@ from scipy import constants
 import numpy as np
 import matplotlib.pyplot as plt
 
-from vivarium.actor.process import Process, deep_merge, convert_to_timeseries, plot_simulation_output
+from vivarium.actor.process import Process, convert_to_timeseries, plot_simulation_output
 from vivarium.utils.units import units
 from vivarium.utils.cobra_fba import CobraFBA
-from vivarium.utils.dict_utils import tuplify_role_dicts
+from vivarium.utils.dict_utils import tuplify_role_dicts, deep_merge
 from vivarium.utils.regulation_logic import build_rule
 
 # external concentrations lower than exchange threshold are considered depleted
