+++ vivarium/utils/dict_utils.py
@@ -10,6 +10,19 @@ def merge_dicts(dicts):
         merge.update(d)
     return merge
 
+def deep_merge(dct, merge_dct):
+    '''
+    Recursive dict merge
+    This mutates dct - the contents of merge_dct are added to dct (which is also returned).
+    If you want to keep dct you could call it like deep_merge(dict(dct), merge_dct)'''
+    for k, v in merge_dct.items():
+        if (k in dct and isinstance(dct[k], dict)
+                and isinstance(merge_dct[k], collections.Mapping)):
+            deep_merge(dct[k], merge_dct[k])
+        else:
+            dct[k] = merge_dct[k]
+    return dct
+
 def flatten_role_dicts(dicts):
     '''
     Input:
